pyproject.toml
[project]
name = "docker-volume-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.90",
    "botocore>=1.35.90",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
]
